+++ src/main/java/kr/hhplus/be/commerce/application/coupon/UserCouponIssueProcessor.java
@@ -2,52 +2,20 @@ package kr.hhplus.be.commerce.application.coupon;
 
 import java.time.LocalDateTime;
 
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.transaction.annotation.Transactional;
-
 import kr.hhplus.be.commerce.domain.coupon.model.Coupon;
 import kr.hhplus.be.commerce.domain.coupon.model.UserCoupon;
-import kr.hhplus.be.commerce.domain.coupon.repository.CouponRepository;
-import kr.hhplus.be.commerce.domain.coupon.repository.UserCouponRepository;
-import kr.hhplus.be.commerce.domain.global.exception.CommerceCode;
-import kr.hhplus.be.commerce.domain.global.exception.CommerceException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
-@RequiredArgsConstructor
-public class UserCouponIssueProcessor {
-	private final CouponRepository couponRepository;
-	private final UserCouponRepository userCouponRepository;
+public interface UserCouponIssueProcessor {
+	Output execute(Command command);
 
-	@Transactional
-	public Output execute(Command command) {
-		Coupon issuedCoupon = couponRepository.findByIdForUpdate(command.couponId)
-			.orElseThrow(() -> new CommerceException(CommerceCode.NOT_FOUND_COUPON))
-			.issue(command.now);
-
-		if (userCouponRepository.existsByUserIdAndCouponId(command.userId, command.couponId)) {
-			throw new CommerceException(CommerceCode.ALREADY_ISSUED_COUPON);
-		}
-
-		try {
-			return new Output(
-				couponRepository.save(issuedCoupon),
-				userCouponRepository.save(UserCoupon.of(command.userId, issuedCoupon, command.now))
-			);
-		} catch (DataIntegrityViolationException e) {
-			throw new CommerceException(CommerceCode.ALREADY_ISSUED_COUPON);
-		}
-	}
-
-	public record Command(
+	record Command(
 		Long userId,
 		Long couponId,
 		LocalDateTime now
 	) {
 	}
 
-	public record Output(
+	record Output(
 		Coupon coupon,
 		UserCoupon userCoupon
 	) {
